refactor, fix tests
